fixing minor bug, fixed tests.

+++ API/SmartPlant/Controllers/PlantController.cs
@@ -19,6 +19,7 @@ using Microsoft.Extensions.Hosting.Internal;
 using Newtonsoft.Json;
 using RestSharp;
 using RestSharp.Serialization.Json;
+using SmartPlant.Models.API_Model.Admin;
 
 namespace SmartPlant.Controllers
 {
@@ -44,68 +45,6 @@ namespace SmartPlant.Controllers
          *   USER ROLE REQUIRED ENDPOINTS
          *             BELOW
          */
-
-        /// <summary>
-        /// Used for testing
-        /// </summary>
-        /// <remarks>
-        /// </remarks>
-        /// <response code="200">Success</response>
-        [HttpPost]
-        [AllowAnonymous]
-        [Route("/api/Plants/TEST/image")]
-        public IActionResult UploadImage([FromBody] PlantImageDto img)
-        {
-            Console.WriteLine($"ClientID : {clientID}");
-
-            var client = new RestClient("https://api.imgur.com/3/image");
-            client.Timeout = -1;
-            var request = new RestRequest(Method.POST);
-            request.AddHeader("Authorization", $"Client-ID {clientID}");
-            request.AlwaysMultipartFormData = true;
-            request.AddParameter("image", img.Base64ImgString);
-            IRestResponse response = client.Execute(request);
-            Console.WriteLine(response.Content);
-            Trace.WriteLine(response.Content);
-
-
-            var data = JsonConvert.DeserializeObject<ImgurApiSuccessResponse>(response.Content);
-
-            Console.WriteLine(data.Data.link);
-            Console.WriteLine(data.Data.deletehash);
-
-
-
-            Console.WriteLine(data);
-
-            return Ok(response.Content);
-
-        }
-
-        /// <summary>
-        /// Used for testing
-        /// </summary>
-        /// <remarks>
-        /// </remarks>
-        /// <response code="200">Success</response>
-        [HttpDelete]
-        [AllowAnonymous]
-        [Route("/api/Plants/TEST/Image")]
-        public IActionResult DeleteImage([FromBody] PlantImageDto img)
-        {
-            Console.WriteLine(clientID);
-
-            var client = new RestClient($"https://api.imgur.com/3/image/{img.Base64ImgString}");
-            client.Timeout = -1;
-            var request = new RestRequest(Method.DELETE);
-            request.AddHeader("Authorization", $"Client-ID {clientID}");
-            request.AlwaysMultipartFormData = true;
-            IRestResponse response = client.Execute(request);
-            Console.WriteLine(response.Content);
-            Trace.WriteLine(response.Content);
-            return Ok(response.Content);
-
-        }
 
 
         /// <summary>
@@ -225,7 +164,7 @@ namespace SmartPlant.Controllers
 
 
         /// <summary>
-        /// Updates a plants name
+        /// Updates a plants name and Image. If Image string is null, it remains the same. If a new image string is passed the old one is deleted from Imgur.
         /// </summary>
         /// <remarks>
         /// This takes in a plant ID and a string name.
@@ -390,6 +329,11 @@ namespace SmartPlant.Controllers
         [Route("/api/Admin/Plants")]
         public async Task<IActionResult> AdminGetAll()
         {
+            if (!await AdminCheck())
+            {
+                return Unauthorized();
+            }
+
             var plants = await _repo.AdminGetAll();
 
             if (plants == null)
@@ -413,6 +357,10 @@ namespace SmartPlant.Controllers
         [Authorize(Roles = UserRoles.Admin)]
         public async Task<IActionResult> AdminGet(string id)
         {
+            if (!await AdminCheck())
+            {
+                return Unauthorized();
+            }
 
             var plants = await _repo.GetAllForUser(id);
 
@@ -440,6 +388,11 @@ namespace SmartPlant.Controllers
         [Route("/api/Admin/Plants")]
         public async Task<IActionResult> AdminPost([FromBody] AdminAddPlantDto plantDto)
         {
+            if (!await AdminCheck())
+            {
+                return Unauthorized();
+            }
+
             var user = await _userManager.FindByIdAsync(plantDto.UserID);
 
             var genericError = new GenericReturnMessageDto();
@@ -506,8 +459,12 @@ namespace SmartPlant.Controllers
         /// <response code="404">Plant Not Found</response>
         [HttpPut]
         [Route("/api/Admin/Plants")]
-        public async Task<IActionResult> AdminUpdate([FromBody] UpdatePlantDto dto)
+        public async Task<IActionResult> AdminUpdate([FromBody] AdminUpdatePlantDto dto)
         {
+            if (!await AdminCheck())
+            {
+                return Unauthorized();
+            }
             //user id is not needed, since this is an admin action the userID is not relevant
             var plant = new Plant() { Name = dto.Name, PlantID = dto.PlantID };
             var result = await _repo.AdminUpdate(plant);
@@ -538,6 +495,10 @@ namespace SmartPlant.Controllers
         [Route("/api/Admin/Plants")]
         public async Task<IActionResult> AdminDelete(string plantID)
         {
+            if (!await AdminCheck())
+            {
+                return Unauthorized();
+            }
             var result = await _repo.AdminDelete(plantID);
 
             if (result)
@@ -567,6 +528,10 @@ namespace SmartPlant.Controllers
         [Route("/api/Admin/Plants/NewToken/{userID}/{plantID}")]
         public async Task<IActionResult> AdminGenerateNewPlantToken(string userID, string plantID)
         {
+            if (!await AdminCheck())
+            {
+                return Unauthorized();
+            }
             var plantToken = GeneratePlantToken(plantID);
             var result = await _repo.AdminGenerateNewPlantToken(userID, plantToken);
 
@@ -590,6 +555,10 @@ namespace SmartPlant.Controllers
         [Route("/api/Admin/Plants/Image")]
         public async Task<IActionResult> AdminDeletePlantImage(string userID, string plantID)
         {
+            if (!await AdminCheck())
+            {
+                return Unauthorized();
+            }
             var result = await _repo.DeletePlantImage(clientID, plantID, userID);
             return Ok(result);
         }
@@ -606,6 +575,16 @@ namespace SmartPlant.Controllers
                 Token = token
             };
             return plantToken;
+        }
+        private async Task<bool> AdminCheck()
+        {
+            var userId = User?.Identity?.Name;
+            var user = await _userManager.FindByIdAsync(userId);
+            if ((await _userManager.GetRolesAsync(user))?[0] != "Admin")
+            {
+                return false;
+            }
+            return true;
         }
     }
 }
